Document public members
