+++ packages/google_maps_routes_api/lib/src/types/polyline.dart
@@ -1,11 +1,19 @@
 import 'location.dart';
 
+/// Encapsulates an encoded [Polyline].
 class Polyline {
+  /// Creates a [Polyline].
   const Polyline({this.encodedPolyline, this.geoJsonLinestring});
 
+  /// The [String] encoding of the [Polyline] using the polyline encoding algorithm.
+  ///
+  /// https://developers.google.com/maps/documentation/utilities/polylinealgorithm
   final String? encodedPolyline;
+
+  /// Specifies a [Polyline] using the [GeoJsonLinestring] format
   final GeoJsonLinestring? geoJsonLinestring;
-  // TODO: Support for GeoJsonLinestring
+
+  /// Decodes a JSON object to a [Polyline].
   static Polyline? fromJson(Object? json) {
     if (json == null) {
       return null;
@@ -20,6 +28,7 @@ class Polyline {
             : null);
   }
 
+  /// Returns a JSON representation of the [Polyline].
   Map<String, dynamic> toJson() {
     final Map<String, dynamic> json = <String, dynamic>{
       'encodedPolyline': encodedPolyline,
@@ -31,12 +40,20 @@ class Polyline {
   }
 }
 
+/// Encapsulates information for a [Polyline] with an array of two or more positions.
+///
+/// https://www.rfc-editor.org/rfc/rfc7946#section-3.1.4
 class GeoJsonLinestring {
+  /// Creates a [GeoJsonLinestring].
   const GeoJsonLinestring({required this.coordinates, required this.type});
 
+  /// The type for [GeoJsonLinestring] is always "LineString".
   final String type;
+
+  /// Array of two or more positions represented as [LatLng] objects.
   final List<LatLng> coordinates;
 
+  /// Decodes a JSON object to a [GeoJsonLinestring].
   static GeoJsonLinestring? fromJson(Object? json) {
     if (json == null) {
       return null;
@@ -52,6 +69,7 @@ class GeoJsonLinestring {
     );
   }
 
+  /// Returns a JSON representation of the [GeoJsonLinestring].
   Map<String, dynamic> toJson() {
     final Map<String, dynamic> json = <String, dynamic>{
       'type': type,
