+++ packages/google_maps_routes_api/lib/src/types/route.dart
@@ -3,8 +3,12 @@ import 'location.dart';
 import 'polyline.dart';
 import 'route_leg.dart';
 import 'travel_advisory.dart';
+import 'waypoint.dart';
 
+/// Encapsulates a [Route], which consists of a series of connected road segments
+/// that join beginning, ending, and intermediate [Waypoint] points.
 class Route {
+  /// Creates a [Route].
   const Route({
     this.routeLabels,
     this.legs,
@@ -19,18 +23,54 @@ class Route {
     this.routeToken,
   });
 
+  /// Labels for the Route that are useful to identify specific properties of
+  /// the route to compare against others.
   final List<RouteLabel>? routeLabels;
+
+  /// A collection of [RouteLeg] objects (path segments between [Waypoint]
+  /// points) that make-up the [Route]. [RouteLeg] corresponds to the trip
+  /// between two non-via [Waypoint] points.
   final List<RouteLeg>? legs;
+
+  /// The travel distance of the [Route], in meters.
   final int? distanceMeters;
+
+  /// The length of time needed to navigate the [Route] while taking traffic
+  /// conditions into consideration.
+  ///
+  /// A duration in seconds with up to nine fractional digits, ending with 's'.
   final String? duration;
+
+  /// The duration of traveling through the [Route] without taking traffic
+  /// conditions into consideration.
+  ///
+  /// A duration in seconds with up to nine fractional digits, ending with 's'.
   final String? staticDuration;
+
+  /// The overall route [Polyline]. This will be the combined [Polyline]
+  /// of all [RouteLeg] objects.
   final Polyline? polyline;
+
+  /// A description of the route.
   final String? description;
+
+  /// An array of warnings to show when displaying the [Route].
   final List<String>? warnings;
+
+  /// The [Viewport] bounding box of the [Polyline].
   final Viewport? viewport;
+
+  /// Additional information of the [Route].
   final RouteTravelAdvisory? travelAdvisory;
+
+  /// Web-safe base64 encoded [Route] token that can be passed to
+  /// NavigationSDK, which allows the Navigation SDK to reconstruct the [Route]
+  /// during navigation, and in the event of rerouting honor the original
+  /// intention when Routes v2.computeRoutes is called. Customers should treat
+  /// this token as an opaque blob.
   final String? routeToken;
 
+  /// Decodes a JSON object to a [Route].
   static Route? fromJson(Object? json) {
     if (json == null) {
       return null;
@@ -39,11 +79,15 @@ class Route {
     final Map<String, dynamic> data = json as Map<String, dynamic>;
     final List<RouteLabel>? routeLabels = data['routeLabels'] != null
         ? List<RouteLabel>.from(
-            data['routeLabels'].map((label) => RouteLabel.values.byName(label)))
+            List<String>.from(data['routeLabels']).map(
+              (String label) => RouteLabel.values.byName(label),
+            ),
+          )
         : null;
 
     final List<RouteLeg> legs = List<RouteLeg>.from(
         data['legs'].map((model) => RouteLeg.fromJson(model)));
+
     return Route(
       routeLabels: routeLabels,
       legs: legs,
@@ -65,6 +109,7 @@ class Route {
     );
   }
 
+  /// Returns a JSON representation of the [Route].
   Map<String, dynamic> toJson() {
     final Map<String, dynamic> json = <String, dynamic>{
       'routeLabels':
@@ -86,32 +131,66 @@ class Route {
   }
 }
 
+/// [Route] objects have a [Viewport] property.
+///
+/// A latitude-longitude [LatLng] viewport, represented as two diagonally
+/// opposite [low] and [high] points. A viewport is considered a
+/// closed region, i.e. it includes its boundary.
+///
+/// The latitude bounds must range between -90 to 90 degrees
+///  inclusive, and the longitude bounds must range between
+/// -180 to 180 degrees inclusive.
+///
+/// * If [low] = [high], the viewport consists of that single point.
+///
+/// * If [low.longitude] > [high.longitude], the longitude range is
+///   inverted (the viewport crosses the 180 degree longitude line).
+///
+/// * If [low.longitude] = -180 degrees and [high.longitude] = 180 degrees,
+///   the viewport includes all longitudes.
+///
+/// * If [low.longitude] = 180 degrees and [high.longitude] = -180 degrees,
+///   the longitude range is empty.
+///
+/// * If [low.latitude] > [high.latitude], the latitude range is empty.
+///
+/// Both [low] and [high] must be populated, and the represented box cannot
+/// be empty (as specified by the definitions above).
 class Viewport {
-  const Viewport({this.low, this.high});
+  /// Creates a [Viewport].
+  const Viewport({required this.low, required this.high});
 
-  final LatLng? low;
-  final LatLng? high;
+  /// The low point of the [Viewport].
+  final LatLng low;
 
+  /// The high point of the [Viewport].
+  final LatLng high;
+
+  /// Decodes a JSON object to a [Viewport].
   static Viewport? fromJson(Object? json) {
     if (json == null) {
       return null;
     }
-    assert(json is Map<String, dynamic>);
-    final Map<String, dynamic> data = json as Map<String, dynamic>;
 
+    assert(json is Map<String, dynamic>);
+
+    final Map<String, dynamic> data = json as Map<String, dynamic>;
+    assert(data['low'] != null);
+    assert(data['high'] != null);
     return Viewport(
-      low: data['low'] ? LatLng.fromMap(data['low']) : null,
-      high: data['high'] ? LatLng.fromMap(data['high']) : null,
+      low: LatLng.fromMap(data['low'])!,
+      high: LatLng.fromMap(data['high'])!,
     );
   }
 
+  /// Returns a JSON representation of the [Viewport].
   Map<String, dynamic> toJson() {
     final Map<String, dynamic> json = <String, dynamic>{
-      'low': low?.toMap(),
-      'high': high?.toMap(),
+      'low': low.toMap(),
+      'high': high.toMap(),
     };
 
-    json.removeWhere((String key, value) => value == null);
+    json.removeWhere((String key, dynamic value) => value == null);
     return json;
   }
 }
