+++ packages/google_maps_routes_api/lib/src/types/route_leg.dart
@@ -1,9 +1,13 @@
 import 'location.dart';
 import 'navigation_instruction.dart';
 import 'polyline.dart';
+import 'routes_request.dart';
 import 'travel_advisory.dart';
+import 'waypoint.dart';
 
+/// Encapsulates a segment between non-via [Waypoint] objects.
 class RouteLeg {
+  /// Creates a [RouteLeg].
   const RouteLeg({
     this.distanceMeters,
     this.duration,
@@ -15,15 +19,45 @@ class RouteLeg {
     this.travelAdvisory,
   });
 
+  /// The travel distance of the [RouteLeg], in meters.
   final int? distanceMeters;
+
+  /// The length of time needed to navigate the [RouteLeg] while taking traffic
+  /// conditions into consideration.
+  ///
+  /// A duration in seconds with up to nine fractional digits, ending with 's'.
   final String? duration;
+
+  /// The duration of traveling through the [RouteLeg] without taking traffic
+  /// conditions into consideration.
+  ///
+  /// A duration in seconds with up to nine fractional digits, ending with 's'.
   final String? staticDuration;
+
+  /// The overall [Polyline] for this [RouteLeg]. This includes the [Polyline]
+  /// for each [RouteLegStep].
   final Polyline? polyline;
+
+  /// The start [Location] of this [RouteLeg]. This might be different from the
+  /// provided [RoutesRequest.origin]. For example, when the provided
+  /// [RoutesRequest.origin] is not near a road, this is a point on the road.
   final Location? startLocation;
+
+  /// The end [Location] of this [RouteLeg]. This might be different from the
+  /// provided [RoutesRequest.destination]. For example, when the provided
+  /// [RoutesRequest.destination] is not near a road, this is a point on the
+  /// road.
   final Location? endLocation;
+
+  /// An array of [RouteLegStep] objects denoting segments within this
+  /// [RouteLeg]. Each [RouteLegStep] represents one [NavigationInstruction].
   final List<RouteLegStep>? steps;
+
+  /// Encapsulates the additional information that the user should be informed
+  /// about, such as possible traffic zone restriction etc. on a [RouteLeg].
   final RouteLegTravelAdvisory? travelAdvisory;
 
+  /// Decodes a JSON object to a [RouteLeg].
   static RouteLeg? fromJson(Object? json) {
     if (json == null) {
       return null;
@@ -51,6 +85,7 @@ class RouteLeg {
     );
   }
 
+  /// Returns a JSON representation of the [RouteLeg].
   Map<String, dynamic> toJson() {
     final Map<String, dynamic> json = <String, dynamic>{
       'distanceMeters': distanceMeters,
@@ -68,7 +103,11 @@ class RouteLeg {
   }
 }
 
+/// Encapsulates a segment of a [RouteLeg]. A [RouteLegStep] corresponds to a
+/// single [NavigationInstruction]. [RouteLeg] objects are made up of
+/// [RouteLegStep] objects.
 class RouteLegStep {
+  /// Creates a [RouteLegStep].
   const RouteLegStep({
     this.distanceMeters,
     this.staticDuration,
@@ -79,14 +118,34 @@ class RouteLegStep {
     this.travelAdvisory,
   });
 
+  /// The travel distance of this [RouteLegStep], in meters.
+  /// In some circumstances, this field might not have a value.
   final int? distanceMeters;
+
+  /// The duration of travel through this [RouteLegStep] without taking traffic
+  /// conditions into consideration. In some circumstances, this field might
+  /// not have a value.
+  ///
+  /// A duration in seconds with up to nine fractional digits, ending with 's'.
   final String? staticDuration;
+
+  /// The [Polyline] associated with this [RouteLegStep].
   final Polyline? polyline;
+
+  /// The start [Location] of this [RouteLegStep].
   final Location? endLocation;
+
+  /// The end [Location] of this [RouteLegStep].
   final Location? startLocation;
+
+  /// Navigation instructions
   final NavigationInstruction? navigationInstruction;
+
+  /// Encapsulates the additional information that the user should be informed
+  /// about, such as possible traffic zone restriction on a [RouteLegStep].
   final RouteLegStepTravelAdvisory? travelAdvisory;
 
+  /// Decodes a JSON object to a [RouteLegStep].
   static RouteLegStep? fromJson(Object? json) {
     if (json == null) {
       return null;
@@ -114,6 +173,7 @@ class RouteLegStep {
     );
   }
 
+  /// Returns a JSON representation of the [RouteLegStep].
   Map<String, dynamic> toJson() {
     final Map<String, dynamic> json = <String, dynamic>{
       'distanceMeters': distanceMeters,
