+++ packages/google_maps_routes_api/lib/src/types/route_matrix_request.dart
@@ -1,5 +1,6 @@
 import 'enums.dart';
 import 'routes_request.dart';
+import 'waypoint.dart';
 
 class RouteMatrixRequest {
   RouteMatrixRequest({
