+++ packages/google_maps_routes_api/lib/src/types/routes_request.dart
@@ -1,5 +1,5 @@
 import 'enums.dart';
-import 'location.dart';
+import 'waypoint.dart';
 
 class RoutesRequest {
   RoutesRequest({
@@ -93,33 +93,5 @@ class VehicleInfo {
 
   Map<String, dynamic> toJson() {
     return <String, dynamic>{'emissionType': emissionType.name};
-  }
-}
-
-class Waypoint {
-  const Waypoint({
-    this.via,
-    this.vehicleStopover,
-    this.sideOfRoad,
-    this.location,
-    this.placeId,
-  });
-
-  final bool? via;
-  final bool? vehicleStopover;
-  final bool? sideOfRoad;
-  final Location? location;
-  final String? placeId;
-
-  Map<String, dynamic> toJson() {
-    final Map<String, dynamic> json = <String, dynamic>{
-      'via': via,
-      'vehicleStopover': vehicleStopover,
-      'sideOfRoad': sideOfRoad,
-      'location': location?.toJson(),
-      'placeId': placeId,
-    };
-    json.removeWhere((String key, value) => value == null);
-    return json;
   }
 }
